// src/app/shared/models/aep.model.ts
export interface AepRecord {
  data: string;
  unidade: string;
  setor: string;
  cargo: string;
  caracterizacaoAtividade: string;
  fonteGeradora: string;
  grupoRisco: string;
  subGrupoRisco: string;
  percentualExposicao: string;
  medidasControle: string;
  probabilidade: number;
  severidade: number;
  nivelRisco: number;
  classificacao: string;
  avaliacaoAmbiental: string;
  planoAcao: string;
  responsavel: string;
  prazo: string;
  status: string;
}